Fixed some errors

+++ src/tools/csg_gmxtopol.cc
@@ -69,7 +69,7 @@ bool GmxTopolApp::EvaluateTopology(Topology *top, Topology *top_ref) {
 void GmxTopolApp::WriteAtoms(ostream &out, Molecule &cg) {
   out << "[atoms]\n";
   out << "; nr type resnr residue atom cgnr charge mass\n";
-  for (int i = 0; i < cg.BeadCount(); ++i) {
+  for (size_t i = 0; i < cg.BeadCount(); ++i) {
     Bead *b = cg.getBead(i);
     out << format("%d %s 1 RES %s %d %f %f\n") % (i + 1) % b->getType() %
                b->getName() % (i + 1) % b->getQ() % b->getMass();
